extract complicated fields; needed to extract fieldGenerator
