+++ packages/graphql/src/fields/list.ts
@@ -33,7 +33,14 @@ export const list = ({
   fmt: string;
   field: ListField;
   config: { rootPath: string; siteLookup: string };
-  fieldData: { sectionFmts: any; templateObjectTypes: any; templatePages: any };
+  fieldData: {
+    sectionFmts: any;
+    templateObjectTypes: any;
+    templatePages: any;
+    templateDataObjectTypes: any;
+    templateFormObjectTypes: any;
+    templateDataInputObjectTypes: any;
+  };
 }) => {
   if (isDocumentListField(field)) {
     return {
